requirements.txt
instagrapi
openai
moviepy
python-dotenv
